#20 - restructure models folder and fix model ref errors
