+++ server/src/models/errors/apiError.ts
@@ -1,4 +1,4 @@
-import { LogService } from "./../logic/services/logging.service";
+import { LogService } from "./../../logic/services/logging.service";
 
 
 export class ApiError {
