+++ server/src/models/index.ts
@@ -1,6 +1,6 @@
-export * from "./apiError";
-export * from "./dbError";
-export * from "./socobouser";
-export * from "./loginResult";
-export * from "./comparePwResult";
-export * from "./extractRequestBodyResult";
+export * from "./errors/apiError";
+export * from "./errors/dbError";
+export * from "./database/socobouser";
+export * from "./results/loginResult";
+export * from "./results/comparePwResult";
+export * from "./results/extractRequestBodyResult";
